wip: more profile tweaks
