+++ packages/app/.env.development
@@ -1,3 +1,3 @@
 PUBLIC_WORKER_URL=https://gh-skyline-dev.battlesquid.workers.dev/
 PUBLIC_WORKER_ENTERPRISE_URL="https://gh-skyline-dev.battlesquid.workers.dev?enterprise=true"
-PUBLIC_APP_NAME="github skyline (dev)"
+PUBLIC_APP_NAME="Github Skyline (dev)"
