+++ packages/app/.env.production
@@ -1,3 +1,3 @@
 PUBLIC_WORKER_URL=https://gh-skyline.battlesquid.workers.dev/
 PUBLIC_WORKER_ENTERPRISE_URL="https://gh-skyline.battlesquid.workers.dev/enterprise?enterprise=true"
-PUBLIC_APP_NAME=github skyline
+PUBLIC_APP_NAME=Github Skyline
