+++ packages/app/src/api/auth.ts
@@ -5,21 +5,21 @@ const TOKEN_KEY = "token";
 const LOGIN_KEY = "login";
 
 export interface UserProfile {
-    name: string | null;
-    avatarUrl?: string;
-    login: string;
+  name: string | null;
+  avatarUrl?: string;
+  login: string;
 }
 
 export const fetchProfile = async (): Promise<UserProfile | null> => {
-    const result = await client.query(UserQuery, {}).toPromise();
-    if (!result.data) {
-        return null;
-    }
-    return {
-        name: result.data.viewer.name,
-        avatarUrl: result.data.viewer.avatarUrl as string, // TODO: make sure default avatars aren't affected by this
-        login: result.data.viewer.login
-    }
+  const result = await client.query(UserQuery, {}).toPromise();
+  if (!result.data) {
+    return null;
+  }
+  return {
+    name: result.data.viewer.name,
+    avatarUrl: result.data.viewer.avatarUrl as string, // TODO: make sure default avatars aren't affected by this
+    login: result.data.viewer.login
+  }
 }
 
 export const getToken = () => localStorage.getItem(TOKEN_KEY);
@@ -67,18 +67,18 @@ export const resolveToken = async (): Promise<string | null> => {
 }
 
 export const logout = () => {
-    localStorage.clear();
-    location.reload();
+  deleteToken();
+  location.reload();
 }
 
 export const loadLogin = async () => {
-    if (localStorage.getItem(LOGIN_KEY) === null) {
-        const profile = await fetchProfile();
-        if (profile === null) {
-            return;
-        }
-        localStorage.setItem(LOGIN_KEY, profile.login);
+  if (localStorage.getItem(LOGIN_KEY) === null) {
+    const profile = await fetchProfile();
+    if (profile === null) {
+      return;
     }
+    localStorage.setItem(LOGIN_KEY, profile.login);
+  }
 }
 
 export const getUsername = () => localStorage.getItem(LOGIN_KEY)
