+++ packages/app/src/api/client.ts
@@ -1,6 +1,6 @@
 import { authExchange } from "@urql/exchange-auth";
 import { cacheExchange, Client, fetchExchange } from "urql";
-import { getToken, logout, resolveToken } from "./auth";
+import { getToken, isAuthenticated, logout, resolveToken } from "./auth";
 
 export const client = new Client({
   url: "https://api.github.com/graphql",
@@ -21,7 +21,9 @@ export const client = new Client({
           return getToken() === null || error.response?.status === 401;
         },
         async refreshAuth() {
-          logout();
+          if (isAuthenticated()) {
+            logout();
+          }
         },
       }
     }),
