+++ packages/app/src/components/profile.tsx
@@ -1,4 +1,4 @@
-import { ActionIcon, Avatar, Flex, Text } from "@mantine/core";
+import { ActionIcon, Avatar, Flex, Paper, Text, useMantineTheme } from "@mantine/core";
 import { IconLogout } from "@tabler/icons-react";
 import { logout, UserProfile } from "../api/auth";
 
@@ -11,18 +11,21 @@ export function Profile(props: ProfileProps) {
         return <></>
     }
     const { name, avatarUrl, login } = props.profile;
+    const theme = useMantineTheme();
     return (
-        <Flex align="center" justify="space-between">
-            <Flex align="center" gap={5}>
-                <Avatar src={avatarUrl} alt={name ?? ""} />
-                <div>
-                    <Text fw="bold">{name}</Text>
-                    <Text>{login}</Text>
-                </div>
+        <Paper p={15} style={{ backgroundColor: theme.colors.dark[8] }}>
+            <Flex align="center" justify="space-between">
+                <Flex align="center" gap={10}>
+                    <Avatar src={avatarUrl} alt={name ?? ""} />
+                    <div>
+                        <Text fw="bold">{name}</Text>
+                        <Text size="sm" c="dimmed">{login}</Text>
+                    </div>
+                </Flex>
+                <ActionIcon onClick={() => logout()} variant="light" aria-label="Logout">
+                    <IconLogout style={{ width: "70%", height: "70%" }} stroke={1.5} />
+                </ActionIcon>
             </Flex>
-            <ActionIcon onClick={() => logout()} variant="filled" aria-label="Logout">
-                <IconLogout style={{ width: "70%", height: "70%" }} stroke={1.5} />
-            </ActionIcon>
-        </Flex>
+        </Paper>
     );
 }
