+++ packages/app/src/components/skyline.tsx
@@ -1,9 +1,10 @@
 import { Bounds, Environment, Grid, OrbitControls } from "@react-three/drei";
 import { Canvas, RenderProps } from "@react-three/fiber";
-import { useMemo, useRef } from "react";
+import { Suspense, useMemo, useRef } from "react";
 import { Group } from "three";
 import { SkylineModel, SkylineModelProps } from "./skyline_model";
 import { useParametersStore } from "../stores";
+import { useMantineTheme } from "@mantine/core";
 
 export type SkylineProps = Omit<SkylineModelProps, "group">
 
@@ -13,17 +14,20 @@ export function Skyline(props: SkylineProps) {
   const group = useRef<Group>(null!);
   const style = useMemo(() => ({}), []);
   const camera = useMemo<RenderProps<HTMLCanvasElement>["camera"]>(() => ({ position: [0, 0, 10], fov: 10 }), []);
+  const theme = useMantineTheme();
 
   return (
     <Canvas
       style={style}
       camera={camera}
       shadows
-      
     >
-      <Bounds fit clip observe margin={1}>
-        <SkylineModel group={group} years={years} />
-      </Bounds>
+      <color attach="background" args={[theme.colors.dark[6]]} />
+      <Suspense fallback={null}>
+        <Bounds fit clip observe margin={1}>
+          <SkylineModel group={group} years={years} />
+        </Bounds>
+      </Suspense>
       <ambientLight intensity={Math.PI / 2} />
       <spotLight castShadow position={[0, 20, 200]} angle={0.50} penumbra={0.1} decay={0.4} intensity={Math.PI} />
       <pointLight castShadow position={[0, 40, 50]} decay={0} intensity={Math.PI} />
